- working with load tests playing with connection pool

+++ src/main/resources/db/migration/V001__create_client_table.sql
@@ -1,5 +1,5 @@
 CREATE TABLE client (
 	id SERIAL,
 	name VARCHAR,
-	value INT
+    value NUMERIC(10,2)
 );
